+++ integration-tests/emulated/tests/people/people-kusama/src/tests/mod.rs
@@ -14,6 +14,7 @@
 // limitations under the License.
 
 mod aliases;
+mod burn;
 mod claim_assets;
 mod governance;
 mod identity;